chore: bump ktor 2.3.0 and okHttp 4.10.0
