+++ gradle/libs.versions.toml
@@ -12,9 +12,9 @@ cryptobox4j = "1.3.0"
 cryptobox-android = "1.1.3-logging"
 javax-crypto = "1.1.0-alpha06"
 android-security = "1.1.0-alpha06"
-ktor = "2.2.2"
+ktor = "2.3.0"
 okio = "3.2.0"
-ok-http = "4.9.3"
+ok-http = "4.10.0"
 mockative = "1.2.6"
 android-work = "2.8.1"
 android-test-runner = "1.5.0"
